# packages/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: CariCash Nova API
  version: 0.2.0
  description: |
    CariCash Nova — Barbados-based mobile payments platform API.

    ## Key Concepts
    - **Jurisdiction**: Barbados. Base currency: BBD. NO FX — cross-currency postings rejected.
    - **Closed-loop wallets**: Customers, agents, and merchants hold balances internally.
    - **Double-entry ledger**: Every transaction posts balanced journal entries (sum DR = sum CR).
    - **Idempotency**: All money-moving endpoints require `idempotency_key`. Retries with same key+payload return cached result; same key+different payload returns 409 DUPLICATE_IDEMPOTENCY_CONFLICT.
    - **Maker-checker**: Reversals, manual adjustments, overdraft facilities, fee/commission changes require dual approval (maker ≠ checker).
    - **Posting via Durable Object**: All money movement routes through a serialized Durable Object per wallet-per-currency.
  contact:
    name: CariCash Engineering
    email: engineering@caricash.bb
  license:
    name: MIT
servers:
  - url: /
    description: Current host
  - url: http://localhost:8787
    description: Local development (wrangler dev)
...
security: []
paths:
...
  /float/top-up:
    post:
      operationId: floatTopUp
      tags: [Float]
      summary: Top-up agent float
      description: |
        Staff deposits float into an agent's `CASH_FLOAT` account.
        Posts a double-entry journal (system suspense -> agent cash float), updates account balances,
        and records an auditable float operation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FloatTopUpRequest'
            example:
              agent_code: AGT001
              amount: '500.00'
              currency: BBD
              staff_id: '01HZ...'
              reason: Branch float replenishment
              reference: BANK-DEP-1289
              idempotency_key: float-topup-001
              correlation_id: '01HZ...'
      responses:
        '201':
          description: Float top-up posted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FloatOperationResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          description: Staff not found or unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Idempotency conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /float/withdrawal:
    post:
      operationId: floatWithdrawal
      tags: [Float]
      summary: Withdraw agent float
      description: |
        Staff withdraws float from an agent's `CASH_FLOAT` account back to system suspense.
        Records a float operation and emits float withdrawal events.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FloatWithdrawalRequest'
            example:
              agent_code: AGT001
              amount: '200.00'
              currency: BBD
              staff_id: '01HZ...'
              reason: End-of-day settlement
              reference: BANK-WDL-778
              idempotency_key: float-withdrawal-001
              correlation_id: '01HZ...'
      responses:
        '201':
          description: Float withdrawal posted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FloatOperationResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          description: Staff not found or unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Idempotency conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /float/{agentCode}/balance:
    get:
      operationId: getAgentFloatBalance
      tags: [Float]
      summary: Get agent float balance
      description: Retrieve current actual and available float balances for the agent.
      parameters:
        - name: agentCode
          in: path
          required: true
          schema:
            type: string
        - name: currency
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/CurrencyCode'
      responses:
        '200':
          description: Agent float balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FloatBalanceResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /float/{agentCode}/history:
    get:
      operationId: getAgentFloatHistory
      tags: [Float]
      summary: Get agent float operation history
      description: List float top-up and withdrawal operations for an agent.
      parameters:
        - name: agentCode
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: Float operation history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FloatHistoryResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================================
  #  TRANSACTIONS
  # ============================================================
...
components:
  securitySchemes:
    CustomerAuth:
      type: http
      scheme: bearer
      description: Bearer token obtained from POST /auth/customer/login
    AgentAuth:
      type: http
      scheme: bearer
      description: Bearer token obtained from POST /auth/agent/login
    MerchantAuth:
      type: http
      scheme: bearer
      description: Bearer token obtained from POST /auth/merchant/login
    StaffAuth:
      type: http
      scheme: bearer
      description: Bearer token obtained from POST /auth/staff/login (OAuth stub)
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: 'Validation failed'
            code: VALIDATION_ERROR
            correlation_id: '01HZ...'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: 'Not found'
            code: NOT_FOUND
            correlation_id: '01HZ...'
...
  schemas:
...
    CurrencyCode:
      type: string
      enum: [BBD, USD]
      description: Supported currency codes. BBD is the base currency.
...
    ErrorCodeEnum:
      type: string
      enum:
        - UNAUTHORIZED
        - FORBIDDEN
        - VALIDATION_ERROR
        - INSUFFICIENT_FUNDS
        - CROSS_CURRENCY_NOT_ALLOWED
        - LIMIT_EXCEEDED
        - KYC_REQUIRED
        - BLOCKLISTED
        - DUPLICATE_IDEMPOTENCY_CONFLICT
        - IDEMPOTENCY_IN_PROGRESS
        - MAKER_CHECKER_REQUIRED
        - NOT_FOUND
        - CONFLICT
        - INTERNAL_ERROR
        - UNBALANCED_JOURNAL
        - JOURNAL_NOT_FOUND
        - ACCOUNT_NOT_FOUND
        - MAKER_CHECKER_VIOLATION
        - APPROVAL_NOT_FOUND
        - APPROVAL_ALREADY_DECIDED
        - INVALID_CREDENTIALS
        - ACCOUNT_LOCKED
        - RATE_LIMIT_EXCEEDED
        - SESSION_EXPIRED
        - ACTOR_NOT_FOUND
        - ACTOR_SUSPENDED
        - DUPLICATE_IDENTIFIER
        - MISSING_REQUIRED_FIELD
        - RECONCILIATION_MISMATCH
        - INTEGRITY_VIOLATION
        - REPAIR_FAILED

    # ---- Error response ----
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message
        code:
          $ref: '#/components/schemas/ErrorCodeEnum'
        name:
          type: string
          description: Error class name
        correlation_id:
          type: string
        details:
          type: object
          additionalProperties: true
        retryable:
          type: boolean
          default: false

    # ---- Auth requests ----
...
    FloatTopUpRequest:
      type: object
      required: [agent_code, amount, staff_id, idempotency_key]
      properties:
        agent_code:
          type: string
        amount:
          type: string
        currency:
          $ref: '#/components/schemas/CurrencyCode'
        staff_id:
          type: string
        reason:
          type: string
        reference:
          type: string
        idempotency_key:
          type: string
        correlation_id:
          type: string
    FloatWithdrawalRequest:
      allOf:
        - $ref: '#/components/schemas/FloatTopUpRequest'
    FloatOperation:
      type: object
      properties:
        id:
          type: string
        agent_actor_id:
          type: string
        agent_account_id:
          type: string
        staff_actor_id:
          type: string
        operation_type:
          type: string
          enum: [TOP_UP, WITHDRAWAL, ADJUSTMENT, CORRECTION]
        amount:
          type: string
        currency:
          $ref: '#/components/schemas/CurrencyCode'
        journal_id:
          type: string
        balance_before:
          type: string
        balance_after:
          type: string
        available_before:
          type: string
        available_after:
          type: string
        requires_approval:
          type: boolean
        reason:
          type: string
        reference:
          type: string
        idempotency_key:
          type: string
        correlation_id:
          type: string
        created_at:
          type: string
          format: date-time
    FloatOperationResponse:
      type: object
      properties:
        operation_id:
          type: string
        journal_id:
          type: string
        agent_id:
          type: string
        agent_code:
          type: string
        amount:
          type: string
        currency:
          $ref: '#/components/schemas/CurrencyCode'
        balance_before:
          type: string
        balance_after:
          type: string
        available_before:
          type: string
        available_after:
          type: string
        correlation_id:
          type: string
    FloatBalanceResponse:
      type: object
      properties:
        agent_id:
          type: string
        agent_code:
          type: string
        currency:
          $ref: '#/components/schemas/CurrencyCode'
        account_id:
          type: string
        actual_balance:
          type: string
        available_balance:
          type: string
        hold_amount:
          type: string
        pending_credits:
          type: string
        correlation_id:
          type: string
    FloatHistoryResponse:
      type: object
      properties:
        agent_id:
          type: string
        agent_code:
          type: string
        operations:
          type: array
          items:
            $ref: '#/components/schemas/FloatOperation'
        count:
          type: integer
        correlation_id:
          type: string
...